PP-3664: fix is read only setting persisting between applications
It looks like attributes assigned to the controller context (`this`) are
somewhow persisted and shared between form sessions. This means that
once it has been set to `true` once it will then be true for future form
sessions even if they are not "read only".

This behaviour should definitely be documented alongside the controller
README to make it clear what might be persisted and what won't be.

As we only ever want the "read only" flag set on a case by case basis
move it from mutating the controller `this` to setting it on the
(presumably ephemeral!) `viewModel` -- note that because we're now
assigning it directly to the template context we no longer need to
access the `page.` data in the templates.