FS-4978: dont use nodemon in production
nodemon is designed to be run locally to watch for file changes and
restart the running process. This behaviour could lead to unpredictable
behaviour when running and has security implications for exucting
modified code.

The `NODE_ENV` is often used by servers and dependencies to gate
behaviour that would only be acceptable in a development environment.
I've not audited what would depend on that in this instance but we don't
want to leave room for vulnerabilities through defauling to open
(development).

Note that this continues the practice of including the NODE_ENV in the
package json. This is to aim for this fix to be as least disruptive as
possible but I don't encourage this.

Ideally the script will just take care of starting the server and the
context it is running in should be responsible for appropriately
configuring the environment.