FS-4946: Create copy of 'standard deploy' workflow
In order to properly prevent outdated deployments in CI using
concurrency groups and matrix jobs for parallel deployments, we will
have to set the matrix in the sub-workflow (deployment) which can't
be done using the shared 'standard deploy' workflow.

Copying it to this repo and pointing the main CI workflow to this copy
means we can then update both to work with the runner and adapter
deployments and use concurrency rules, as we have on pre-award and
data-store repos.

These edits will come in subsequent commits.