FS-4946: Update deploy workflow to use matrix strategy
We can now have parallel deployments for each environment and protect
the deployments using concurrency groups in the calling workflow, in the
same setup as the pre-award and data-store workflows.

A number of steps and variables from the original copied deploy workflow
are redundant and have been removed.

As the production deployment currently only deploys one application
(form runner), it continues to use the shared 'standard deploy' workflow
rather than the new 'deploy' with matrix strategy, but this can be updated
once Form Designer is in production.