FS-4978: Install nodemon only for local dev
Even though the `yarn runner production` command does not explicitly
use nodemon to watch files, it's still getting installed in the images
used in deployed environments. To cut down on the risk of hot reloading
of deployed services, this should only install nodemon if the Dockerfile
is being built for local development. `INSTALL_NODEMON` is set to `true`
in a separate PR on the docker runner repo, and explicitly set to `false`
in the github workflow building images for deployement.