FS-4978: Update Runner dockerfile permissions
Now with a dev-specific Dockerfile for local development we can update
the permissions in the production Dockerfile to be more restrictive.

This commit creates a new `runuser` with much fewer permissions and
runs the container process as that user. The original appuser
(now `builduser``) is still used to build the app and install the
dependencies.

Permissions are left largely untouched in the dev Dockerfile to
facilitate local development, but some overly permissive commands
have been changed.