Confirm GPL license for tinymce
We see this warning in the browser console for tinymce because we
haven't confirmed our license type:

TinyMCE is running in evaluation mode. Provide a valid license key or
add license_key: 'gpl' to the init config to agree to the open source
license terms.

We don't have a commercial license, so I've been told to confirm that
GPL is suitable for our needs.