FS-4978: Extend permissions in dev dockerfiles
In the previous commit we restricted the permissions in the two dev
dockerfiles, but this led to issues running the containers which we
didn't have before when all files and directories were copied and owned
by the root user/main appuser with all the permissions.

This adds back more permissions to key yarn files for local development
to work.