FS-4978: Update main docker container directory structure
Based on what was learnt in the previous commits relating to the dev
dockerfiles, the dockerfiles now need more explicit directory structures
in order to work when we reduce the permission levels of the runuser.

This commit allows the .yarn directory to build and execute properly
but still restricts the permissions of the runuser.