FSPT-342 - Disable JWT auth for Dev and Test envs as well as UAT
We are planning to shut down the UAT environment to save on costs. FAB in Production currently uses Form Runner in UAT for previewing forms. We are now switching Prod FAB to use Test Runner instead of UAT Runner, so that Prod FAB is unaffected when UAT is shut down. However, in order to do this we need to make sure that the /publish endpoint is unprotected by JWT auth in Test, as it has been in UAT, because otherwise we cannot preview from Prod FAB. This is because Prod FAB uses Prod Authenticator which authenticates only within the Prod env, not across envs.

This is intended as a temporary measure to unblock UAT environment shutdown. Longer-term we plan to link Prod FAB to Prod Runner. This will allow this environment-conditional disabling of JWT auth to be rightfully abandoned, but requires some thought around ensuring that preview forms don't contaminate the production repository.