FSPT-342 - Use basic auth when publishing from Designer to Runner
The Designer makes a backend call to the Runner - specifically to its /publish endpoint, to publish forms. This is the data persistence mechanism in Designer, and is absolutely key to its functionality. We need to make sure that we are passing basic auth credentials as part of this backend request.

We implement this in line with existing precedents in Pre-Award and Post-Award - we hard-code the basic auth credentials in a URL we store in AWS Parameter Store (e.g., http://username:password@example.com) and then import that variable as a secret that can be leveraged with no change required to application code. This is especially important here because the backend request made by Designer is implemented within the git submodule / base repo digital-form-builder, increasing the complexity of any application code changes.