levellingup->communities domain changes
I've not touched the manifest here so that we are still associating the
deployed app with both levellingup and communities domains. This is
ostensibly for backwards compatibility and anyone who's eg bookmarked
pages; however given this is for form designer and application
questions, we could probably be much more aggressive with sunsetting
the domains on this app specifically than with eg the pre- or post-award
apps.