Fix next page for multi input pages
The call to `getNext` was being passed the current page's post data,
rather than the form session state, meaning that conditions weren't able
to read the correct data to evaluate as true.

I've tested this locally but don't feel confident writing a good test in
any meaningful amount of time. We'll have to test this thoroughly
ourselves instead.