FLS-1452 - Refactor AdapterCacheService to reduce duplicate code
Currently in AdapterCacheService we have a lot of duplicate code, specifically around (A) supporting both in-memory caching of forms configurations, as well as caching via Redis, and (B) Redis client configuration for both forms configurations and session management.

This PR addresses this duplication. In the case of (A), we create a new StorageAdapter interface that abstracts the common caching operations get, set and getKeys, with form keys stored in an in-memory list in the case of the in-memory cache to compensate for the fact that Catbox does not offer out-of-the-box getKeys functionality like Redis. In the case of (B), we create a new utility function createRedisClient.