FLS-1452 - No more loading of forms from fsd_config into cache on app startup
We are going to fetch forms from the Pre-Award API instead. I wanted this to be a cleanup step further down the line but because it relies on setFormConfiguration in the AdapterCacheService that we need to remove, best to just cut all of this stuff out now to simplify development.