FLS-1452 - Refine logging in AdapterCacheService
Logging has become noisy, both in the code and in the console. This commit reduces logging code by passing a string not an object to logger calls, and reduces logging noise in the console by switching away from using the request logger, which introduces the 'req' object into every log message, which is noisy and duplicative, especially given that we log the request path upstream.