Bugfix - Designer - removing checkbox to make field optional from client-side file upload component configuration
We currently have both a 'Make Client side field upload field optional' checkbox and a 'Minimum required files count' number field in the client-side file upload component configuration form. Having both creates ambiguity e.g., if we tick the checkbox to make it optional, but then set minimum required files to 1, or if we leave the checkbox unticked and set minimum required files to 0. Besides, on the Form Runner side we ignore the 'required' field (set by the checkbox) anyway, instead using the 'Minimum required files count'.

In this commit we simply remove the checkbox from the Designer interface, to remove that ambiguity.