Bugfix - Runner - fixing client-side file upload component validation in preview mode
Currently if you have a client-side file upload component with minimumRequiredFiles > 0, it is impossible to succesfully proceed past component validation when you are running the form in preview mode. This is because when we are looking for files to see if the minimumRequiredFiles threshold has been met, we build an S3 path with a form_session_identifier that doesn't exist because we're trying to take it from state.metadata, when in preview mode we need to take it from request.query (form_session_identifier is sent in the URL as one of the query parameters).

This change adjusts client-side file upload component validation to fall back to taking form_session_identifier from the URL in cases where it's not available in the state.metadata, supporting both normal and preview workflows.