FLS-1450 - Handling display name properly inc. modification via 'Form details' form
The issue we had was that in the 'Form details' form, powered by the FormDetails.tsx component, we populate the 'Title' field from the 'name' attribute in the form JSON configuration. And when the form is submitted, it upserts 'name' into the JSON configuration. This is problematic for a couple of reasons. Firstly, 'name' isn't part of the form configuration by default - when you create a new form, it's not there. This is an issue you can observe on main branch - if you create a new form and open the 'Form details' form, you'll see the 'Title' field as blank. And secondly, we don't want to have 'name' in the form JSON configuration anyway - we now store forms as form_definition records in the Pre-Award database with separate display_name and draft_json attributes, and we don't want two references to form name that we need to keep in sync. Thirdly, display names used everywhere - in FAB, in Form Designer - now come from the display_name attribute of the form_definition table, so any changes made via 'Form details' the 'name' attribute of the JSON configuration do not actually do anything.

This commit fixes that, by getting the display name from Pre-Award, passing it through into the data context so we can show in the 'Title' field, and then when the form is submitted, saving the updated title into the React state and also via the Designer API to the Pre-Award API.