FLS-1196 - Fixing failing Form Designer E2E test
The issue is that the Form Designer only has access to the preview namespace of the Runner cache, so simply loading test.json into the cache on startup doesn't suffice because that puts it in the permanent namespace. The fix is to actively publish test.json via the publish API to ensure it's in the preview namespace.