FLS-1561 - Disable JWT auth for Runner for Designer-specific endpoints
The Designer-specific endpoints `GET /published` and `GET /published/{id}` are used specifically by the Form Designer, and will be removed soon. In the case of `GET /published`, Designer uses this to get a list of all forms to display on the `/app/choose-existing` page; in the case of `GET /published/{id}`, Designer uses to load form JSON configurations into memory so that it can render and manipulate them in edit mode.

Why are we disabling JWT auth for these endpoints now? It's because we are moving Form Designer into Production, and that means we are going to bump up against JWT auth on these endpoints for the first time, which Form Designer is not equipped to handle correctly. Form Designer currently makes server-side requests to the endpoints `POST /publish`, `GET /published` and `GET /published/{id}`, and because of the nature of server-side requests, users' valid JWT auth tokens are not passed automatically, and so the solution is for the server to explicitly forward them. It is fiddly to implement this for one server-side request, let alone three, and while we need to retain the `POST /publish` call in the new Pre-Award API-integrated world for the purpose of supporting form previews until Form Runner is also integrated with the Pre-Award API, in the case of `GET /published` and `GET /published/{id}`, these endpoints are on the precipice of becoming redundant in the next few days, as Form Designer pivots to getting lists of forms and individual form JSON configurations from the Pre-Award API. Thus it is not considered a valuable use of time to implement JWT auth forwarding for these two `GET` endpoints, and thus we need to disable JWT auth on the Runner side while we deploy Form Designer into Production, to ensure service continuity.