FLS-1456 - Using matrix deployment of both Designer and Runner in Production
Dev, Test and UAT share a common, custom workflow for deploying both Designer and Runner simultaneously. Previously Designer wasn't deployed in Production so we overrode this with special Production logic to use the standard-deploy workflow from the funding-service-design-workflows repository. Now however we are moving Designer into Production as well, and this means we can share the common, custom workflow in Production.