FLS-1634 - Create four new forms for Crash Test Dummy Fund Round 2
These four forms are designed to feel like a natural representation of the sorts of application forms we deal with in Funding Service, and more importantly to cover key Form Runner functionality in a non-redundant way, for the purpose of optimising E2E tests. A wide range of field types are incorporated, including custom field types like client-side file upload and multi-input; there are examples of complex conditional branching with nested conditions; and we apply data validation throughout including custom validation parameters additional to field defaults.